Fix possible 'race condition' when populating slots after a "new array" execution
It might happen that an array reset occurs before the "retain" list is
created, resulting in an empty list.

There is one open point:
When the form is submitted and the page returns then $var['mdState']
variable isn't updated yet by emhttp, this only occurs after a full page
reload or page switch. As a result the GUI doesn't show the correct
message after pressing Apply.

Is there a way to let emhttp update the $var variable before returning
the page?