Protect GUI from invalid .page files
If the format of the .page is completely wrong, then every page gets corrupted, and the system becomes unusable.  Personally, I hate it that every .page file can potentially take down the entire GUI in case of an error.  Blocking the output of the parse_ini_file protects the GUI, and the file gets logged to inform the user / forum.