Include links in email and Discord agent notifications
Users with the Discord agent currently installed will need to go to the Notification Agents page, make a small change to the Discord agent, and hit Apply to regenerate the script.

Note: When clicking links from web-based email clients you will likely be redirected to the login screen due to SameSite cookie restrictions. Links can be copy/pasted to the address bar instead.