Fix php warning when saving text that contains a reserved php character to a file.
A php warning occurs when a text area is to be saved to a file and it contains a reserved php character.  Update.php tries to parse the file and extract a specific section even if the file is only to be saved.  It is also trying to extract a section when the $section is not defined i.e. $section is set to false.