Update update.wireguard.php
The MTU needs to be set to at least to the wireguard default. While it seems to work just fine, to create a docker bridge with a MTU of 1500, it breaks in some cases. For example, if you "curl https://api.minecraftservices.com/publickeys" via a container thats running over that wireguard-docker-bridge, it fails.